+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `get_M()` function to handle M-variable access regardless of step order.
 - Metadata with detailed metrics information based on each instance solution
 - Graph visualizations on information of each instance.
+- Instance-specific metrics collection for visualization and metadata storage.
 
 ### Changed
 - Segregated out and specified our current implementation of ORTools as UDPB (User-Dependent Pseudo-Boolean).